+++ server/src/weidu-d/provider.ts
@@ -114,7 +114,7 @@ class WeiduDProvider implements LanguageProvider {
             conlog("WeiDU D provider not initialized, cannot compile");
             return;
         }
-        weiduCompile(uri, this.storedContext.settings.weidu, interactive, text);
+        await weiduCompile(uri, this.storedContext.settings.weidu, interactive, text);
     }
 }